Moving Layer to Layers class
